PSPDFKit 2.14.6 - Easier way to change the link border color, faster thumbnail loading for the scrobble bar, improved annotation resize handling when zoomed in, exposed PSPDFRenderStatusView, improved bounding box calculation for line endings, faster search and even better text extraction/word detection and bug fixes.

+++ PSPDFKit.embeddedframework/PSPDFKit.framework/Versions/A/Headers/PSPDFPageRenderer.h
@@ -44,7 +44,7 @@ extern Class PSPDFPageRendererClass;
 - (CGRect)renderPageRef:(CGPDFPageRef)page inContext:(CGContextRef)context inRectangle:(CGRect)rectangle pageInfo:(PSPDFPageInfo *)pageInfo withAnnotations:(NSArray *)annotations options:(NSDictionary *)options;
 
 /// Render a page; defined by point and zoom. Use zoom=100 and point = CGPointMake(0, 0) for defaults.
-- (CGSize)renderPage:(CGPDFPageRef)page inContext:(CGContextRef)context atPoint:(CGPoint)point withZoom:(float)zoom pageInfo:(PSPDFPageInfo *)pageInfo withAnnotations:(NSArray *)annotations options:(NSDictionary *)options;
+- (CGSize)renderPage:(CGPDFPageRef)page inContext:(CGContextRef)context atPoint:(CGPoint)point withZoom:(double)zoom pageInfo:(PSPDFPageInfo *)pageInfo withAnnotations:(NSArray *)annotations options:(NSDictionary *)options;
 
 /// Renders a particular appearance stream (A PDF within a PDF) into a context.
 /// Will return NO if rendering failed.
